moving kubernetes_monitor.py back to runner subdir
