+++ tests/pytest_tests/unit_tests/test_launch/test_runner/test_safe_watch.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock
 import pytest
 from kubernetes.client import ApiException
 from urllib3.exceptions import ProtocolError
-from wandb.sdk.launch.monitor.kubernetes_monitor import SafeWatch
+from wandb.sdk.launch.runner.kubernetes_monitor import SafeWatch
 
 
 class MockWatch:
